Split actions into multiple files, add JSDoc to all action creators, thunks and helpers
