+++ eisbuk-admin/src/enums/store.ts
