+++ eisbuk-admin/src/pages/CustomersPage/CustomersPage.tsx
@@ -13,7 +13,10 @@ import AppbarAdmin from "@/components/layout/AppbarAdmin";
 import CustomerList from "@/components/customers/CustomerList";
 import CustomerForm from "@/components/customers/CustomerForm";
 
-import { deleteCustomer, updateCustomer } from "@/store/actions/firestore";
+import {
+  deleteCustomer,
+  updateCustomer,
+} from "@/store/actions/customerOperations";
 
 import { useTitle } from "@/utils/helpers";
