+++ eisbuk-admin/src/pages/DashboardPage/DashboardPage.tsx
@@ -9,7 +9,7 @@ import DateNavigationAppBar from "@/containers/DateNavigationAppBar";
 import BookingsByDay from "@/components/BookingsByDay";
 import AppbarAdmin from "@/components/layout/AppbarAdmin";
 
-import { markAbsentee } from "@/store/actions/actions";
+import { markAbsentee } from "@/store/actions/bookingOperations";
 import { bookingDayInfoSelector, calendarDaySelector } from "@/store/selectors";
 
 import { useTitle } from "@/utils/helpers";
