+++ eisbuk-admin/src/store/actions/copyPaste.ts
@@ -1,6 +1,6 @@
 import { Slot } from "eisbuk-shared";
 
-import { Action } from "@/enums/Redux";
+import { Action } from "@/enums/store";
 
 import { SlotDay, SlotWeek } from "@/types/store";
